minor fixes

+++ backend/chds_food_delivery/apps/orders/views.py
@@ -117,7 +117,7 @@ class OrderCreateView(APIView):
                 else:
                     transaction = self._make_transaction(order,"STRIPE")
                     payment_intent = stripe.PaymentIntent.create(
-                        amount=(float(order.total_price) * 100),
+                        amount=int(float(order.total_price) * 100),
                         currency='aud',
                         payment_method_types=["card", "alipay", "wechat_pay"],
                         metadata={'order_id': order.order_id, 'user_id': order.user.id},
